Post one toot at a time, save lastProcessedPostId after each one
⚠️ Untested change ⚠️ 

This jumped out at me as a potential problem if there were a number of toots to be cross-posted in one go. Firing them all off at once means that they could potentially end up on bluesky out-of-order.

Additionally, if any failed to cross-post, saving after each successful one would allow the script to pick up where it left off.

I was thinking about changing the script to copy over all of my toots, not just future ones, but I realized that just writing a 1 to the lastProcessedPostId.txt would do that, and then this change would ensure it went in order and was recoverable.

I'll try it out sometime soon - I was going to just file a bug report, but I figured it'd be easier to make the change than to describe it.